HHH-9330 : Test cases for orphanRemoval=true does not work in bidirectional relationships (without cascading)
