+++ hibernate-entitymanager/src/test/java/org/hibernate/ejb/test/orphan/onetomany/Product.java
@@ -21,32 +21,29 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.hibernate.jpa.test.orphan.onetomany;
+package org.hibernate.ejb.test.orphan.onetomany;
 
+import java.util.ArrayList;
+import java.util.List;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 @Entity
-@Table(name="feature")
-public class Feature {
-	@Id @GeneratedValue(strategy= GenerationType.AUTO)
+@Table(name="product")
+public class Product {
+	@Id
+	@GeneratedValue(strategy= GenerationType.AUTO)
 	private Long id;
 
-	@ManyToOne()
-	private Product product;
+	@OneToMany(mappedBy = "product", orphanRemoval = true)
+	//, cascade ={CascadeType.PERSIST}
+	private List<Feature> features = new ArrayList<Feature>();
 
 	private String name;
-
-	public Feature() {
-	}
-
-	public Feature(Product product) {
-		this.product = product;
-	}
 
 	public Long getId() {
 		return id;
@@ -62,11 +59,11 @@ public class Feature {
 		this.name = name;
 	}
 
-	public Product getProduct() {
-		return product;
+	public List<Feature> getFeatures() {
+		return features;
 	}
-	public void setProduct(Product product) {
-		this.product = product;
+	public void setFeatures(List<Feature> features) {
+		this.features = features;
 	}
 }
